split up cw contract
