+++ programs/cw-ics08-wasm-eth/src/contract.rs
@@ -1,8 +1,6 @@
-//! This module handles the execution logic of the contract.
+//! This module contains the `CosmWasm` entrypoints for the 08-wasm smart contract
 
-use cosmwasm_std::{
-    entry_point, to_json_binary, Binary, Deps, DepsMut, Env, MessageInfo, Response,
-};
+use cosmwasm_std::{entry_point, Binary, Deps, DepsMut, Env, MessageInfo, Response};
 use ethereum_light_client::{
     client_state::ClientState as EthClientState,
     consensus_state::ConsensusState as EthConsensusState,
@@ -14,12 +12,12 @@ use ibc_proto::ibc::{
     },
 };
 
-use crate::ContractError;
-use crate::{custom_query::EthereumCustomQuery, state::store_client_state};
+use crate::{custom_query::EthereumCustomQuery, query, state::store_client_state};
 use crate::{
     msg::{ExecuteMsg, InstantiateMsg, QueryMsg, SudoMsg},
     state::store_consensus_state,
 };
+use crate::{sudo, ContractError};
 
 /// The instantiate entry point for the CosmWasm contract.
 /// # Errors
@@ -83,124 +81,6 @@ pub fn sudo(
     Ok(Response::default().set_data(result))
 }
 
-mod sudo {
-    use cosmwasm_std::DepsMut;
-    use ethereum_light_client::update::update_consensus_state;
-    use ibc_proto::ibc::{
-        core::client::v1::Height as IbcProtoHeight,
-        lightclients::wasm::v1::ConsensusState as WasmConsensusState,
-    };
-
-    use crate::{
-        msg::{
-            Height, UpdateStateMsg, UpdateStateResult, VerifyMembershipMsg, VerifyNonMembershipMsg,
-        },
-        state::{
-            get_eth_client_state, get_eth_consensus_state, get_wasm_client_state,
-            store_client_state, store_consensus_state,
-        },
-    };
-
-    use super::{to_json_binary, Binary, ContractError, Deps, EthereumCustomQuery};
-
-    pub fn verify_membership(
-        deps: Deps<EthereumCustomQuery>,
-        verify_membership_msg: VerifyMembershipMsg,
-    ) -> Result<Binary, ContractError> {
-        let eth_client_state = get_eth_client_state(deps.storage);
-        let eth_consensus_state =
-            get_eth_consensus_state(deps.storage, verify_membership_msg.height.revision_height);
-
-        ethereum_light_client::membership::verify_membership(
-            eth_consensus_state,
-            eth_client_state,
-            verify_membership_msg.proof.into(),
-            verify_membership_msg
-                .merkle_path
-                .key_path
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-            Some(verify_membership_msg.value.into()),
-        )
-        .map_err(ContractError::VerifyMembershipFailed)?;
-
-        Ok(Binary::default())
-    }
-
-    pub fn verify_non_membership(
-        deps: Deps<EthereumCustomQuery>,
-        verify_non_membership_msg: VerifyNonMembershipMsg,
-    ) -> Result<Binary, ContractError> {
-        let eth_client_state = get_eth_client_state(deps.storage);
-        let eth_consensus_state = get_eth_consensus_state(
-            deps.storage,
-            verify_non_membership_msg.height.revision_height,
-        );
-
-        ethereum_light_client::membership::verify_membership(
-            eth_consensus_state,
-            eth_client_state,
-            verify_non_membership_msg.proof.into(),
-            verify_non_membership_msg
-                .merkle_path
-                .key_path
-                .into_iter()
-                .map(Into::into)
-                .collect(),
-            None,
-        )
-        .map_err(ContractError::VerifyNonMembershipFailed)?;
-
-        Ok(Binary::default())
-    }
-
-    #[allow(clippy::needless_pass_by_value)]
-    pub fn update_state(
-        deps: DepsMut<EthereumCustomQuery>,
-        update_state_msg: UpdateStateMsg,
-    ) -> Result<Binary, ContractError> {
-        let header_bz: Vec<u8> = update_state_msg.client_message.into();
-        let header = serde_json::from_slice(&header_bz)
-            .map_err(ContractError::DeserializeClientMessageFailed)?;
-
-        let eth_client_state = get_eth_client_state(deps.storage);
-        let eth_consensus_state =
-            get_eth_consensus_state(deps.storage, eth_client_state.latest_slot);
-
-        let (updated_slot, updated_consensus_state, updated_client_state) =
-            update_consensus_state(&eth_consensus_state, &eth_client_state, header)
-                .map_err(ContractError::UpdateClientStateFailed)?;
-
-        let consensus_state_bz: Vec<u8> = serde_json::to_vec(&updated_consensus_state)
-            .map_err(ContractError::SerializeConsensusStateFailed)?;
-        let wasm_consensus_state = WasmConsensusState {
-            data: consensus_state_bz,
-        };
-        store_consensus_state(deps.storage, &wasm_consensus_state, updated_slot)?;
-
-        if let Some(client_state) = updated_client_state {
-            let client_state_bz: Vec<u8> = serde_json::to_vec(&client_state)
-                .map_err(ContractError::SerializeClientStateFailed)?;
-
-            let mut wasm_client_state = get_wasm_client_state(deps.storage);
-            wasm_client_state.data = client_state_bz;
-            wasm_client_state.latest_height = Some(IbcProtoHeight {
-                revision_number: 0,
-                revision_height: updated_slot,
-            });
-            store_client_state(deps.storage, &wasm_client_state)?;
-        }
-
-        Ok(to_json_binary(&UpdateStateResult {
-            heights: vec![Height {
-                revision_number: 0,
-                revision_height: updated_slot,
-            }],
-        })?)
-    }
-}
-
 /// Execute entry point is not used in this contract.
 #[entry_point]
 #[allow(clippy::needless_pass_by_value, clippy::missing_errors_doc)]
@@ -227,86 +107,16 @@ pub fn query(
         QueryMsg::VerifyClientMessage(verify_client_message_msg) => {
             query::verify_client_message(deps, env, verify_client_message_msg)
         }
-        QueryMsg::CheckForMisbehaviour(_) => query::check_for_misbehaviour(),
-        QueryMsg::TimestampAtHeight(_) => query::timestamp_at_height(env),
+        QueryMsg::CheckForMisbehaviour(_) => todo!(),
+        QueryMsg::TimestampAtHeight(timestamp_at_height_msg) => {
+            query::timestamp_at_height(deps, timestamp_at_height_msg)
+        }
         QueryMsg::Status(_) => query::status(),
-    }
-}
-
-mod query {
-    use crate::{
-        custom_query::BlsVerifier,
-        msg::{
-            CheckForMisbehaviourResult, StatusResult, TimestampAtHeightResult,
-            VerifyClientMessageMsg,
-        },
-        state::{get_eth_client_state, get_eth_consensus_state},
-    };
-
-    use super::{to_json_binary, Binary, ContractError, Deps, Env, EthereumCustomQuery};
-
-    #[allow(clippy::needless_pass_by_value)]
-    pub fn verify_client_message(
-        deps: Deps<EthereumCustomQuery>,
-        env: Env,
-        verify_client_message_msg: VerifyClientMessageMsg,
-    ) -> Result<Binary, ContractError> {
-        let eth_client_state = get_eth_client_state(deps.storage);
-        let eth_consensus_state =
-            get_eth_consensus_state(deps.storage, eth_client_state.latest_slot);
-        let header = serde_json::from_slice(&verify_client_message_msg.client_message)
-            .map_err(ContractError::DeserializeClientMessageFailed)?;
-        let bls_verifier = BlsVerifier {
-            querier: deps.querier,
-        };
-
-        ethereum_light_client::verify::verify_header(
-            &eth_consensus_state,
-            &eth_client_state,
-            env.block.time.seconds(),
-            &header,
-            bls_verifier,
-        )
-        .map_err(ContractError::VerifyClientMessageFailed)?;
-
-        Ok(Binary::default())
-    }
-
-    pub fn check_for_misbehaviour() -> Result<Binary, ContractError> {
-        Ok(to_json_binary(&CheckForMisbehaviourResult {
-            found_misbehaviour: false,
-        })?)
-    }
-
-    #[allow(clippy::needless_pass_by_value)]
-    pub fn timestamp_at_height(env: Env) -> Result<Binary, ContractError> {
-        let now = env.block.time.seconds();
-        Ok(to_json_binary(&TimestampAtHeightResult { timestamp: now })?)
-    }
-
-    pub fn status() -> Result<Binary, ContractError> {
-        Ok(to_json_binary(&StatusResult {
-            status: "Active".to_string(),
-        })?)
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use std::marker::PhantomData;
-
-    use alloy_primitives::B256;
-    use cosmwasm_std::{
-        testing::{
-            mock_dependencies, MockApi, MockQuerier, MockQuerierCustomHandlerResult, MockStorage,
-        },
-        Binary, OwnedDeps, SystemResult,
-    };
-    use ethereum_light_client::test_utils::bls_verifier::{aggreagate, fast_aggregate_verify};
-    use ethereum_types::consensus::bls::{BlsPublicKey, BlsSignature};
-
-    use crate::custom_query::EthereumCustomQuery;
-
     mod instantiate_tests {
         use alloy_primitives::{Address, FixedBytes, B256, U256};
         use cosmwasm_std::{
@@ -328,9 +138,10 @@ mod tests {
         use prost::{Message, Name};
 
         use crate::{
-            contract::{instantiate, tests::mk_deps},
+            contract::instantiate,
             msg::InstantiateMsg,
             state::{consensus_db_key, HOST_CLIENT_STATE_KEY},
+            test::mk_deps,
         };
 
         #[test]
@@ -431,45 +242,95 @@ mod tests {
         }
     }
 
-    mod sudo_tests {
+    mod integration_tests {
         use cosmwasm_std::{
-            coins, from_json,
+            coins,
             testing::{message_info, mock_env},
-            Binary,
+            Binary, Timestamp,
         };
         use ethereum_light_client::test_utils::fixtures::{
             self, CommitmentProof, InitialState, StepsFixture, UpdateClient,
         };
 
         use crate::{
-            contract::{instantiate, sudo, tests::mk_deps},
+            contract::{instantiate, query, sudo},
             msg::{
-                Height, MerklePath, SudoMsg, UpdateStateMsg, UpdateStateResult, VerifyMembershipMsg,
+                Height, MerklePath, QueryMsg, SudoMsg, UpdateStateMsg, UpdateStateResult,
+                VerifyClientMessageMsg, VerifyMembershipMsg,
             },
+            test::mk_deps,
         };
 
         #[test]
-        fn test_verify_membership() {
+        // This test runs throught the e2e test scenario defined in the interchaintest:
+        // TestICS20TransferERC20TokenfromEthereumToCosmosAndBack_Groth16
+        fn test_ics20_transfer_from_ethereum_to_cosmos_flow() {
             let mut deps = mk_deps();
             let creator = deps.api.addr_make("creator");
             let info = message_info(&creator, &coins(1, "uatom"));
 
             let fixture: StepsFixture =
-                fixtures::load("TestICS20TransferNativeCosmosCoinsToEthereumAndBack_Groth16");
+                fixtures::load("TestICS20TransferERC20TokenfromEthereumToCosmosAndBack_Groth16");
 
-            let commitment_proof_fixture: CommitmentProof = fixture.get_data_at_step(2);
+            let initial_state: InitialState = fixture.get_data_at_step(0);
 
-            let client_state = commitment_proof_fixture.client_state;
+            let client_state = initial_state.client_state;
+            let consensus_state = initial_state.consensus_state;
+
             let client_state_bz: Vec<u8> = serde_json::to_vec(&client_state).unwrap();
-            let consensus_state = commitment_proof_fixture.consensus_state;
             let consensus_state_bz: Vec<u8> = serde_json::to_vec(&consensus_state).unwrap();
 
-            let msg = crate::msg::InstantiateMsg {
+            let instantiate_msg = crate::msg::InstantiateMsg {
                 client_state: Binary::from(client_state_bz),
                 consensus_state: Binary::from(consensus_state_bz),
-                checksum: b"checksum".into(), // TODO: Real checksum important?
+                checksum: b"checksum".into(),
             };
-            instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+            instantiate(deps.as_mut(), mock_env(), info, instantiate_msg).unwrap();
+
+            // At this point, the light clients are initialized and the client state is stored
+            // In the flow, an ICS20 transfer has been initiated from Ethereum to Cosmos
+            // Next up we want to prove the packet on the Cosmos chain, so we start by updating the
+            // light client (which is two steps: verify client message and update state)
+
+            // Verify client message
+            let update_client: UpdateClient = fixture.get_data_at_step(1);
+            assert_eq!(1, update_client.updates.len()); // just to make sure
+            let header = update_client.updates[0].clone();
+            let header_bz: Vec<u8> = serde_json::to_vec(&header).unwrap();
+
+            // We update the enviornment to be after finalized execution timestamp
+            let mut env = mock_env();
+            env.block.time = Timestamp::from_seconds(
+                header.consensus_update.attested_header.execution.timestamp + 1000,
+            );
+
+            let query_verify_client_msg = QueryMsg::VerifyClientMessage(VerifyClientMessageMsg {
+                client_message: Binary::from(header_bz.clone()),
+            });
+            query(deps.as_ref(), env.clone(), query_verify_client_msg).unwrap();
+
+            // Update state
+
+            let sudo_update_state_msg = SudoMsg::UpdateState(UpdateStateMsg {
+                client_message: Binary::from(header_bz),
+            });
+            let update_res = sudo(deps.as_mut(), env.clone(), sudo_update_state_msg).unwrap();
+            let update_state_result: UpdateStateResult =
+                serde_json::from_slice(&update_res.data.unwrap())
+                    .expect("update state result should be deserializable");
+            assert_eq!(1, update_state_result.heights.len());
+            assert_eq!(0, update_state_result.heights[0].revision_number);
+            assert_eq!(
+                header.consensus_update.attested_header.beacon.slot,
+                update_state_result.heights[0].revision_height
+            );
+
+            // The client has now been updated, and we would submit the packet to the cosmos chain,
+            // along with the proof of th packet commitment. IBC will call verify_membership.
+
+            // Verify memebership
+            let commitment_proof_fixture: CommitmentProof = fixture.get_data_at_step(2);
 
             let proof = commitment_proof_fixture.storage_proof;
             let proof_bz = serde_json::to_vec(&proof).unwrap();
@@ -477,7 +338,7 @@ mod tests {
             let value = proof.value;
             let value_bz = value.to_be_bytes_vec();
 
-            let msg = SudoMsg::VerifyMembership(VerifyMembershipMsg {
+            let query_verify_membership_msg = SudoMsg::VerifyMembership(VerifyMembershipMsg {
                 height: Height {
                     revision_number: 0,
                     revision_height: commitment_proof_fixture.proof_slot,
@@ -490,211 +351,7 @@ mod tests {
                 },
                 value: Binary::from(value_bz),
             });
-            let res = sudo(deps.as_mut(), mock_env(), msg).unwrap();
-            assert_eq!(0, res.messages.len());
-        }
-
-        #[test]
-        fn test_update_state() {
-            let mut deps = mk_deps();
-            let creator = deps.api.addr_make("creator");
-            let info = message_info(&creator, &coins(1, "uatom"));
-
-            let fixture: StepsFixture =
-                fixtures::load("TestICS20TransferNativeCosmosCoinsToEthereumAndBack_Groth16");
-
-            let initial_state: InitialState = fixture.get_data_at_step(0);
-
-            let client_state = initial_state.client_state;
-
-            let consensus_state = initial_state.consensus_state;
-
-            let client_state_bz: Vec<u8> = serde_json::to_vec(&client_state).unwrap();
-            let consensus_state_bz: Vec<u8> = serde_json::to_vec(&consensus_state).unwrap();
-
-            let msg = crate::msg::InstantiateMsg {
-                client_state: Binary::from(client_state_bz),
-                consensus_state: Binary::from(consensus_state_bz),
-                checksum: b"checksum".into(), // TODO: Real checksum important?
-            };
-
-            instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-            let update_client: UpdateClient = fixture.get_data_at_step(1);
-            let header = update_client.updates[0].clone();
-            let header_bz: Vec<u8> = serde_json::to_vec(&header).unwrap();
-
-            let msg = SudoMsg::UpdateState(UpdateStateMsg {
-                client_message: Binary::from(header_bz),
-            });
-            let res = sudo(deps.as_mut(), mock_env(), msg).unwrap();
-            let data = res.data.unwrap();
-            let update_state_result: UpdateStateResult = from_json(data).unwrap();
-            assert_eq!(1, update_state_result.heights.len());
-            assert_eq!(
-                header.consensus_update.attested_header.beacon.slot,
-                update_state_result.heights[0].revision_height
-            );
-        }
-    }
-
-    mod query_tests {
-        use cosmwasm_std::{
-            coins, from_json,
-            testing::{message_info, mock_env},
-            Binary, Timestamp,
-        };
-        use ethereum_light_client::test_utils::fixtures::{
-            self, InitialState, StepsFixture, UpdateClient,
-        };
-
-        use crate::{
-            contract::{instantiate, query, tests::mk_deps},
-            msg::{
-                CheckForMisbehaviourMsg, CheckForMisbehaviourResult, Height, QueryMsg, StatusMsg,
-                StatusResult, TimestampAtHeightMsg, TimestampAtHeightResult,
-                VerifyClientMessageMsg,
-            },
-        };
-
-        #[test]
-        fn test_verify_client_message() {
-            let mut deps = mk_deps();
-            let creator = deps.api.addr_make("creator");
-            let info = message_info(&creator, &coins(1, "uatom"));
-
-            let fixture: StepsFixture =
-                fixtures::load("TestICS20TransferNativeCosmosCoinsToEthereumAndBack_Groth16");
-
-            let initial_state: InitialState = fixture.get_data_at_step(0);
-
-            let client_state = initial_state.client_state;
-
-            let consensus_state = initial_state.consensus_state;
-
-            let client_state_bz: Vec<u8> = serde_json::to_vec(&client_state).unwrap();
-            let consensus_state_bz: Vec<u8> = serde_json::to_vec(&consensus_state).unwrap();
-
-            let msg = crate::msg::InstantiateMsg {
-                client_state: Binary::from(client_state_bz),
-                consensus_state: Binary::from(consensus_state_bz),
-                checksum: b"checksum".into(), // TODO: Real checksum important?
-            };
-
-            instantiate(deps.as_mut(), mock_env(), info, msg).unwrap();
-
-            let update_client: UpdateClient = fixture.get_data_at_step(1);
-            let header = update_client.updates[0].clone();
-            let header_bz: Vec<u8> = serde_json::to_vec(&header).unwrap();
-
-            let mut env = mock_env();
-            env.block.time = Timestamp::from_seconds(
-                header.consensus_update.attested_header.execution.timestamp + 1000,
-            );
-
-            query(
-                deps.as_ref(),
-                env,
-                QueryMsg::VerifyClientMessage(VerifyClientMessageMsg {
-                    client_message: Binary::from(header_bz),
-                }),
-            )
-            .unwrap();
-        }
-
-        #[test]
-        fn test_check_for_misbehaviour() {
-            let deps = mk_deps();
-            let res = query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::CheckForMisbehaviour(CheckForMisbehaviourMsg {
-                    client_message: Binary::default(),
-                }),
-            )
-            .unwrap();
-            let misbehaviour_result: CheckForMisbehaviourResult = from_json(&res).unwrap();
-            assert!(!misbehaviour_result.found_misbehaviour);
-        }
-
-        #[test]
-        fn test_timestamp_at_height() {
-            let deps = mk_deps();
-            let res = query(
-                deps.as_ref(),
-                mock_env(),
-                QueryMsg::TimestampAtHeight(TimestampAtHeightMsg {
-                    height: Height {
-                        revision_number: 0,
-                        revision_height: 1,
-                    },
-                }),
-            )
-            .unwrap();
-            let timestamp_at_height_result: TimestampAtHeightResult = from_json(&res).unwrap();
-            assert_eq!(
-                mock_env().block.time.seconds(),
-                timestamp_at_height_result.timestamp
-            );
-        }
-
-        #[test]
-        fn test_status() {
-            let deps = mk_deps();
-            let res = query(deps.as_ref(), mock_env(), QueryMsg::Status(StatusMsg {})).unwrap();
-            let status_response: StatusResult = from_json(&res).unwrap();
-            assert_eq!("Active", status_response.status);
-        }
-    }
-
-    pub fn custom_query_handler(query: &EthereumCustomQuery) -> MockQuerierCustomHandlerResult {
-        match query {
-            EthereumCustomQuery::AggregateVerify {
-                public_keys,
-                message,
-                signature,
-            } => {
-                let public_keys = public_keys
-                    .iter()
-                    .map(|pk| pk.as_ref().try_into().unwrap())
-                    .collect::<Vec<&BlsPublicKey>>();
-                let message = B256::try_from(message.as_slice()).unwrap();
-                let signature = BlsSignature::try_from(signature.as_slice()).unwrap();
-
-                fast_aggregate_verify(&public_keys, message, signature).unwrap();
-
-                SystemResult::Ok(cosmwasm_std::ContractResult::Ok::<Binary>(
-                    serde_json::to_vec(&true).unwrap().into(),
-                ))
-            }
-            EthereumCustomQuery::Aggregate { public_keys } => {
-                let public_keys = public_keys
-                    .iter()
-                    .map(|pk| pk.as_ref().try_into().unwrap())
-                    .collect::<Vec<&BlsPublicKey>>();
-
-                let aggregate_pubkey = aggreagate(&public_keys).unwrap();
-
-                SystemResult::Ok(cosmwasm_std::ContractResult::Ok::<Binary>(
-                    serde_json::to_vec(&Binary::from(aggregate_pubkey.as_slice()))
-                        .unwrap()
-                        .into(),
-                ))
-            }
-        }
-    }
-
-    fn mk_deps(
-    ) -> OwnedDeps<MockStorage, MockApi, MockQuerier<EthereumCustomQuery>, EthereumCustomQuery>
-    {
-        let deps = mock_dependencies();
-
-        OwnedDeps {
-            storage: deps.storage,
-            api: deps.api,
-            querier: MockQuerier::<EthereumCustomQuery>::new(&[])
-                .with_custom_handler(custom_query_handler),
-            custom_query_type: PhantomData,
+            sudo(deps.as_mut(), env, query_verify_membership_msg).unwrap();
         }
     }
 }
