+++ programs/cw-ics08-wasm-eth/src/lib.rs
@@ -11,6 +11,11 @@ pub mod contract;
 pub mod custom_query;
 mod error;
 pub mod msg;
+pub mod query;
 pub mod state;
+pub mod sudo;
 
 pub use error::ContractError;
+
+#[cfg(test)]
+mod test;
